ChatBackend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.2.0
channels-redis==4.0
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
pillow==11.0.0
psycopg==3.2.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
service-identity==24.2.0
setuptools==75.5.0
sqlparse==0.5.2
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
zope.interface==7.1.1
